@@ -1,130 +1,122 @@
 {
   "name": "docsify",
-  "version": "4.12.2",
+  "version": "4.13.0",
   "description": "A magical documentation generator.",
-  "author": {
-    "name": "qingwei-li",
-    "email": "cinwell.li@gmail.com",
-    "url": "https://github.com/QingWei-Li"
-  },
   "homepage": "https://docsify.js.org",
+  "repository": "github:docsifyjs/docsify",
+  "authors": "https://github.com/docsifyjs/docsify/graphs/contributors",
   "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/docsifyjs/docsify.git"
+  "collective": {
+    "url": "https://opencollective.com/docsify"
   },
-  "main": "lib/docsify.js",
-  "unpkg": "lib/docsify.min.js",
+  "keywords": [
+    "client",
+    "creator",
+    "crs",
+    "doc",
+    "docs",
+    "documentation",
+    "generator",
+    "markdown"
+  ],
+  "engines": {
+    "node": ">=20.11.0"
+  },
+  "type": "module",
+  "main": "dist/docsify.js",
+  "exports": {
+    ".": "./src/core/Docsify.js",
+    "./*": "./*"
+  },
   "files": [
-    "lib",
+    "dist",
     "themes"
   ],
-  "scripts": {
-    "bootstrap": "npm i && lerna bootstrap && npm run build:ssr",
-    "serve": "node server",
-    "serve:ssr": "cross-env SSR=1 node server",
-    "dev": "run-p serve watch:*",
-    "dev:ssr": "run-p serve:ssr watch:*",
-    "lint": "eslint .",
-    "fixlint": "eslint . --fix",
-    "test": "jest",
-    "build:test": "npm run build && npm test",
-    "test:e2e": "jest --selectProjects e2e",
-    "test:integration": "jest --selectProjects integration",
-    "test:unit": "jest --selectProjects unit",
-    "css": "node build/css",
-    "watch:css": "npm run css -- -o themes -w",
-    "watch:js": "node build/build.js",
-    "build:css:min": "mkdirp lib/themes && npm run css -- -o lib/themes && node build/mincss.js",
-    "build:css": "mkdirp themes && npm run css -- -o themes",
-    "build:js": "cross-env NODE_ENV=production node build/build.js",
-    "build:ssr": "node build/ssr.js",
-    "build:cover": "node build/cover.js",
-    "build": "rimraf lib themes && run-s build:js build:css build:css:min build:ssr build:cover",
-    "prepare": "npm run build",
-    "pub:next": "cross-env RELEASE_TAG=next sh build/release.sh",
-    "pub": "sh build/release.sh",
-    "postinstall": "opencollective-postinstall",
-    "docker:build": "docker build -f Dockerfile -t docsify-test:local .",
-    "docker:clean": "docker rmi docsify-test:local",
-    "docker:rebuild": "npm run docker:clean && npm run docker:build",
-    "docker:test": "npm run docker:cli -- test",
-    "docker:build:test": "npm run docker:cli -- build:test",
-    "docker:test:e2e": "npm run docker:cli -- test:e2e",
-    "docker:test:integration": "npm run docker:cli -- test:integration",
-    "docker:test:unit": "npm run docker:cli -- test:unit",
-    "docker:cli": "docker run --rm -it --ipc=host --mount type=bind,source=$(pwd)/test,target=/app/test docsify-test:local"
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
-  },
   "lint-staged": {
     "*.js": "eslint --fix"
   },
   "dependencies": {
-    "dompurify": "^2.3.1",
-    "marked": "^1.2.9",
-    "medium-zoom": "^1.0.6",
+    "dexie": "^4.0.10",
+    "medium-zoom": "^1.1.0",
     "opencollective-postinstall": "^2.0.2",
-    "prismjs": "^1.23.0",
-    "strip-indent": "^3.0.0",
-    "tinydate": "^1.3.0",
-    "tweezer.js": "^1.4.0"
+    "prismjs": "^1.29.0",
+    "tinydate": "^1.3.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.11.6",
+    "@babel/eslint-parser": "^7.24.5",
     "@babel/preset-env": "^7.11.5",
-    "autoprefixer-stylus": "^1.0.0",
-    "axios": "^0.21.1",
-    "babel-eslint": "^10.0.3",
-    "babel-jest": "^26.3.0",
-    "browser-sync": "^2.26.12",
-    "chokidar": "^3.4.2",
+    "@eslint/js": "^10.0.0",
+    "@playwright/test": "^1.44.0",
+    "@rollup/plugin-babel": "^6.0.4",
+    "@rollup/plugin-commonjs": "^28.0.1",
+    "@rollup/plugin-node-resolve": "^16.0.0",
+    "@rollup/plugin-replace": "^6.0.1",
+    "@rollup/plugin-terser": "^0.4.3",
+    "@types/eslint": "^8.40.2",
+    "axios": "^1.5.0",
+    "browser-sync": "^3.0.2",
     "common-tags": "^1.8.0",
-    "conventional-changelog-cli": "^2.1.0",
-    "copy-dir": "^1.2.0",
-    "cross-env": "^6.0.3",
-    "cssnano": "^4.1.10",
-    "eslint": "^5.16.0",
-    "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-jest": "^24.0.2",
-    "eslint-plugin-jest-playwright": "^0.2.1",
-    "eslint-plugin-prettier": "^3.1.2",
-    "husky": "^3.1.0",
-    "jest": "^26.4.2",
-    "jest-image-snapshot": "^4.2.0",
-    "jest-playwright-preset": "^1.4.1",
-    "lerna": "^3.22.1",
-    "lint-staged": "^10.4.0",
-    "live-server": "^1.2.1",
-    "mkdirp": "^0.5.1",
+    "conventional-changelog-cli": "^3.0.0",
+    "cross-env": "^7.0.3",
+    "cssnano": "^7.0.1",
+    "eslint": "^9.3.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-jest": "^28.5.0",
+    "eslint-plugin-playwright": "^2.1.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "glob": "^10.3.15",
+    "globals": "^15.3.0",
+    "husky": "^9.0.11",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "lint-staged": "^15.2.2",
+    "marked": "^15.0.3",
     "npm-run-all": "^4.1.5",
-    "playwright": "^1.8.0",
-    "prettier": "^1.19.1",
-    "rimraf": "^3.0.0",
-    "rollup": "^1.23.1",
-    "rollup-plugin-async": "^1.2.0",
-    "rollup-plugin-buble": "^0.19.8",
-    "rollup-plugin-commonjs": "^10.1.0",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-replace": "^2.2.0",
-    "rollup-plugin-uglify": "^6.0.4",
-    "serve-handler": "^6.1.2",
-    "stylus": "^0.54.5",
-    "vue2": "npm:vue@^2.6.12",
-    "vue3": "npm:vue@^3.0.0",
+    "postcss-cli": "^11.0.0",
+    "postcss-import": "^16.1.0",
+    "postcss-nesting": "^13.0.0",
+    "prettier": "^3.2.5",
+    "rimraf": "^5.0.7",
+    "rollup": "^4.17.2",
+    "rollup-plugin-import-css": "^3.5.0",
+    "vue": "^3.4.27",
     "xhr-mock": "^2.5.1"
   },
-  "keywords": [
-    "doc",
-    "docs",
-    "documentation",
-    "creator",
-    "generator"
-  ],
-  "collective": {
-    "url": "https://opencollective.com/docsify"
+  "scripts": {
+    "build:cover": "node build/cover.js",
+    "build:css": "postcss \"src/themes/*.css\" \"src/themes/**/[!_]*.css\" --base src/themes --dir dist/themes --map",
+    "build:css:min": "cross-env NODE_ENV='production' npm run build:css -- --ext .min.css",
+    "build:emoji": "node ./build/emoji.js",
+    "build:js": "rollup -c",
+    "build": "run-s clean build:js build:css build:css:min build:cover",
+    "clean": "rimraf --glob dist/** themes/** _playwright*/**",
+    "dev": "run-p serve:dev watch:*",
+    "docker:build:test": "npm run docker:cli -- build:test",
+    "docker:build": "docker build -f Dockerfile -t docsify-test:local .",
+    "docker:clean": "docker rmi docsify-test:local",
+    "docker:cli": "docker run --rm -it --ipc=host --mount type=bind,source=$(pwd)/test,target=/app/test docsify-test:local",
+    "docker:rebuild": "run-s docker:clean docker:build",
+    "docker:test:e2e": "npm run docker:cli -- test:e2e",
+    "docker:test:integration": "npm run docker:cli -- test:integration",
+    "docker:test:unit": "npm run docker:cli -- test:unit",
+    "docker:test": "npm run docker:cli -- test",
+    "lint:fix": "prettier . --write && eslint . --fix",
+    "lint": "prettier . --check && eslint .",
+    "postinstall": "opencollective-postinstall && npx husky install",
+    "prepare": "npm run build",
+    "prettier": "prettier . --write",
+    "pub:next": "cross-env RELEASE_TAG=next sh build/release.sh",
+    "pub": "sh build/release.sh",
+    "serve:dev": "npm run serve -- --dev",
+    "serve": "node server",
+    "test:e2e": "playwright test",
+    "test:e2e:chromium": "playwright test --project='chromium'",
+    "test:e2e:ui": "playwright test --ui",
+    "test:integration": "npm run test:jest -- --selectProjects integration",
+    "test:jest": "cross-env NODE_OPTIONS=--experimental-vm-modules jest",
+    "test:unit": "npm run test:jest -- --selectProjects unit",
+    "test": "run-s test:jest test:e2e",
+    "watch:css": "run-p 'build:css -- --watch' 'build:css:min -- --watch'",
+    "watch:js": "npm run build:js -- --watch"
   }
 }
