@@ -1,28 +1,28 @@
-const docsifyInit = require('../helpers/docsify-init');
+import docsifyInit from '../helpers/docsify-init.js';
+import { test, expect } from './fixtures/docsify-init-fixture.js';
 
-describe(`Index file hosting`, function() {
+test.describe('Index file hosting', () => {
   const sharedOptions = {
     config: {
-      basePath: `${TEST_HOST}/docs/index.html#/`,
+      basePath: '/index.html#/',
     },
-    testURL: `${TEST_HOST}/docs/index.html#/`,
+    testURL: '/index.html#/',
   };
 
-  test('should serve from index file', async () => {
+  test('should serve from index file', async ({ page }) => {
     await docsifyInit(sharedOptions);
-
-    await expect(page).toHaveText(
-      '#main',
-      'A magical documentation site generator'
+    await expect(page.locator('#main')).toContainText(
+      'A magical documentation site generator',
     );
     expect(page.url()).toMatch(/index\.html#\/$/);
   });
 
-  test('should use index file links in sidebar from index file hosting', async () => {
+  test('should use index file links in sidebar from index file hosting', async ({
+    page,
+  }) => {
     await docsifyInit(sharedOptions);
-
     await page.click('a[href="#/quickstart"]');
-    await expect(page).toHaveText('#main', 'Quick start');
+    await expect(page.locator('#main')).toContainText('Quick start');
     expect(page.url()).toMatch(/index\.html#\/quickstart$/);
   });
 });
