@@ -1,6 +1,7 @@
-const docsifyInit = require('../helpers/docsify-init');
+import docsifyInit from '../helpers/docsify-init.js';
+import { test, expect } from './fixtures/docsify-init-fixture.js';
 
-describe(`Security`, function() {
+test.describe('Security - Cross Site Scripting (XSS)', () => {
   const sharedOptions = {
     markdown: {
       homepage: '# Hello World',
@@ -9,24 +10,25 @@ describe(`Security`, function() {
       'test.md': '# Test Page',
     },
   };
+  const slashStrings = ['//', '///'];
 
-  describe(`Cross Site Scripting (XSS)`, function() {
-    const slashStrings = ['//', '///'];
+  for (const slashString of slashStrings) {
+    const hash = `#${slashString}domain.com/file.md`;
 
-    for (const slashString of slashStrings) {
-      const hash = `#${slashString}domain.com/file.md`;
+    test(`should not load remote content from hash (${hash})`, async ({
+      page,
+    }) => {
+      const mainElm = page.locator('#main');
 
-      test(`should not load remote content from hash (${hash})`, async () => {
-        await docsifyInit(sharedOptions);
-        await expect(page).toHaveText('#main', 'Hello World');
-        await page.evaluate(() => (location.hash = '#/test'));
-        await expect(page).toHaveText('#main', 'Test Page');
-        await page.evaluate(newHash => {
-          location.hash = newHash;
-        }, hash);
-        await expect(page).toHaveText('#main', 'Hello World');
-        expect(page.url()).toMatch(/#\/$/);
-      });
-    }
-  });
+      await docsifyInit(sharedOptions);
+      await expect(mainElm).toContainText('Hello World');
+      await page.evaluate(() => (location.hash = '#/test'));
+      await expect(mainElm).toContainText('Test Page');
+      await page.evaluate(newHash => {
+        location.hash = newHash;
+      }, hash);
+      await expect(mainElm).toContainText('Hello World');
+      expect(page.url()).toMatch(/#\/$/);
+    });
+  }
 });
