@@ -1,11 +1,10 @@
-const {
+import {
   removeAtag,
   getAndRemoveConfig,
-} = require('../../src/core/render/utils');
-
-const { tree } = require(`../../src/core/render/tpl`);
-
-const { slugify } = require(`../../src/core/render/slugify`);
+  getAndRemoveDocsifyIgnoreConfig,
+} from '../../src/core/render/utils.js';
+import { tree } from '../../src/core/render/tpl.js';
+import { slugify } from '../../src/core/render/slugify.js';
 
 // Suite
 // -----------------------------------------------------------------------------
@@ -16,7 +15,47 @@ describe('core/render/utils', () => {
     test('removeAtag from a link', () => {
       const result = removeAtag('<a href="www.example.com">content</a>');
 
-      expect(result).toEqual('content');
+      expect(result).toBe('content');
+    });
+  });
+
+  // getAndRemoveDocsifyIgnoreConfig()
+  // ---------------------------------------------------------------------------
+  describe('getAndRemoveDocsifyIgnoreConfig()', () => {
+    test('getAndRemoveDocsifyIgnoreConfig from <!-- {docsify-ignore} -->', () => {
+      const { content, ignoreAllSubs, ignoreSubHeading } =
+        getAndRemoveDocsifyIgnoreConfig(
+          'My Ignore Title<!-- {docsify-ignore} -->',
+        );
+      expect(content).toBe('My Ignore Title');
+      expect(ignoreSubHeading).toBeTruthy();
+      expect(ignoreAllSubs === undefined).toBeTruthy();
+    });
+
+    test('getAndRemoveDocsifyIgnoreConfig from <!-- {docsify-ignore-all} -->', () => {
+      const { content, ignoreAllSubs, ignoreSubHeading } =
+        getAndRemoveDocsifyIgnoreConfig(
+          'My Ignore Title<!-- {docsify-ignore-all} -->',
+        );
+      expect(content).toBe('My Ignore Title');
+      expect(ignoreAllSubs).toBeTruthy();
+      expect(ignoreSubHeading === undefined).toBeTruthy();
+    });
+
+    test('getAndRemoveDocsifyIgnoreConfig from {docsify-ignore}', () => {
+      const { content, ignoreAllSubs, ignoreSubHeading } =
+        getAndRemoveDocsifyIgnoreConfig('My Ignore Title{docsify-ignore}');
+      expect(content).toBe('My Ignore Title');
+      expect(ignoreSubHeading).toBeTruthy();
+      expect(ignoreAllSubs === undefined).toBeTruthy();
+    });
+
+    test('getAndRemoveDocsifyIgnoreConfig from {docsify-ignore-all}', () => {
+      const { content, ignoreAllSubs, ignoreSubHeading } =
+        getAndRemoveDocsifyIgnoreConfig('My Ignore Title{docsify-ignore-all}');
+      expect(content).toBe('My Ignore Title');
+      expect(ignoreAllSubs).toBeTruthy();
+      expect(ignoreSubHeading === undefined).toBeTruthy();
     });
   });
 
@@ -25,18 +64,18 @@ describe('core/render/utils', () => {
   describe('getAndRemoveConfig()', () => {
     test('parse simple config', () => {
       const result = getAndRemoveConfig(
-        `[filename](_media/example.md ':include')`
+        "[filename](_media/example.md ':include')",
       );
 
       expect(result).toMatchObject({
         config: {},
-        str: `[filename](_media/example.md ':include')`,
+        str: "[filename](_media/example.md ':include')",
       });
     });
 
     test('parse config with arguments', () => {
       const result = getAndRemoveConfig(
-        `[filename](_media/example.md ':include :foo=bar :baz test')`
+        "[filename](_media/example.md ':include :foo=bar :baz test')",
       );
 
       expect(result).toMatchObject({
@@ -44,18 +83,18 @@ describe('core/render/utils', () => {
           foo: 'bar',
           baz: true,
         },
-        str: `[filename](_media/example.md ':include test')`,
+        str: "[filename](_media/example.md ':include test')",
       });
     });
 
     test('parse config with double quotes', () => {
       const result = getAndRemoveConfig(
-        `[filename](_media/example.md ":include")`
+        '[filename](_media/example.md ":include")',
       );
 
       expect(result).toMatchObject({
         config: {},
-        str: `[filename](_media/example.md ":include")`,
+        str: '[filename](_media/example.md ":include")',
       });
     });
   });
@@ -82,8 +121,8 @@ describe('core/render/tpl', () => {
       },
     ]);
 
-    expect(result).toEqual(
-      `<ul class="app-sub-sidebar"><li><a class="section-link" href="#/cover?id=basic-usage" title="Basic usage"><span style="color:red">Basic usage</span></a></li><li><a class="section-link" href="#/cover?id=custom-background" title="Custom background">Custom background</a></li><li><a class="section-link" href="#/cover?id=test" title="Test"><img src="/docs/_media/favicon.ico" data-origin="/_media/favicon.ico" alt="ico">Test</a></li></ul>`
+    expect(result).toBe(
+      /* html */ '<ul class="app-sub-sidebar"><li><a class="section-link" href="#/cover?id=basic-usage" title="Basic usage"><span style="color:red">Basic usage</span></a></li><li><a class="section-link" href="#/cover?id=custom-background" title="Custom background">Custom background</a></li><li><a class="section-link" href="#/cover?id=test" title="Test"><img src="/docs/_media/favicon.ico" data-origin="/_media/favicon.ico" alt="ico">Test</a></li></ul>',
     );
   });
 });
@@ -91,12 +130,12 @@ describe('core/render/tpl', () => {
 describe('core/render/slugify', () => {
   test('slugify()', () => {
     const result = slugify(
-      `Bla bla bla <svg aria-label="broken" class="broken" viewPort="0 0 1 1"><circle cx="0.5" cy="0.5"/></svg>`
+      'Bla bla bla <svg aria-label="broken" class="broken" viewPort="0 0 1 1"><circle cx="0.5" cy="0.5"/></svg>',
     );
     const result2 = slugify(
-      `Another <span style="font-size: 1.2em" class="foo bar baz">broken <span class="aaa">example</span></span>`
+      'Another <span style="font-size: 1.2em" class="foo bar baz">broken <span class="aaa">example</span></span>',
     );
-    expect(result).toEqual(`bla-bla-bla-`);
-    expect(result2).toEqual(`another-broken-example`);
+    expect(result).toBe('bla-bla-bla-');
+    expect(result2).toBe('another-broken-example');
   });
 });
