@@ -1,4 +1,4 @@
-const { resolvePath } = require('../../src/core/util');
+import { resolvePath } from '../../src/core/util/index.js';
 
 // Suite
 // -----------------------------------------------------------------------------
@@ -9,19 +9,19 @@ describe('router/util', () => {
     test('resolvePath with filename', () => {
       const result = resolvePath('hello.md');
 
-      expect(result).toEqual('/hello.md');
+      expect(result).toBe('/hello.md');
     });
 
     test('resolvePath with ./', () => {
       const result = resolvePath('./hello.md');
 
-      expect(result).toEqual('/hello.md');
+      expect(result).toBe('/hello.md');
     });
 
     test('resolvePath with ../', () => {
       const result = resolvePath('test/../hello.md');
 
-      expect(result).toEqual('/hello.md');
+      expect(result).toBe('/hello.md');
     });
   });
 });
