@@ -2,20 +2,21 @@ import { Router } from './router/index.js';
 import { Render } from './render/index.js';
 import { Fetch } from './fetch/index.js';
 import { Events } from './event/index.js';
-import initGlobalAPI from './global-api.js';
+import { VirtualRoutes } from './virtual-routes/index.js';
 
 import config from './config.js';
-import { isFn } from './util/core';
-import { Lifecycle } from './init/lifecycle';
+import { isFn } from './util/core.js';
+import { Lifecycle } from './init/lifecycle.js';
 
 /** @typedef {new (...args: any[]) => any} Constructor */
 
-// eslint-disable-next-line new-cap
-export class Docsify extends Fetch(Events(Render(Router(Lifecycle(Object))))) {
+export class Docsify extends Fetch(
+  Events(Render(VirtualRoutes(Router(Lifecycle(Object))))),
+) {
+  config = config(this);
+
   constructor() {
     super();
-
-    this.config = config(this);
 
     this.initLifecycle(); // Init hooks
     this.initPlugin(); // Install plugins
@@ -28,13 +29,19 @@ export class Docsify extends Fetch(Events(Render(Router(Lifecycle(Object))))) {
   }
 
   initPlugin() {
-    []
-      .concat(this.config.plugins)
-      .forEach(fn => isFn(fn) && fn(this._lifecycle, this));
+    this.config.plugins.forEach(fn => {
+      try {
+        isFn(fn) && fn(this._lifecycle, this);
+      } catch (err) {
+        if (this.config.catchPluginErrors) {
+          const errTitle = 'Docsify plugin error';
+
+          // eslint-disable-next-line no-console
+          console.error(errTitle, err);
+        } else {
+          throw err;
+        }
+      }
+    });
   }
 }
-
-/**
- * Global API
- */
-initGlobalAPI();
