@@ -1,6 +1,6 @@
-import { noop } from '../util/core';
+import { noop } from '../util/core.js';
 
-/** @typedef {import('../Docsify').Constructor} Constructor */
+/** @typedef {import('../Docsify.js').Constructor} Constructor */
 
 /**
  * @template {!Constructor} T
@@ -8,6 +8,9 @@ import { noop } from '../util/core';
  */
 export function Lifecycle(Base) {
   return class Lifecycle extends Base {
+    _hooks = {};
+    _lifecycle = {};
+
     initLifecycle() {
       const hooks = [
         'init',
@@ -18,9 +21,6 @@ export function Lifecycle(Base) {
         'ready',
       ];
 
-      this._hooks = {};
-      this._lifecycle = {};
-
       hooks.forEach(hook => {
         const arr = (this._hooks[hook] = []);
         this._lifecycle[hook] = fn => arr.push(fn);
@@ -29,22 +29,48 @@ export function Lifecycle(Base) {
 
     callHook(hookName, data, next = noop) {
       const queue = this._hooks[hookName];
+      const catchPluginErrors = this.config.catchPluginErrors;
 
-      const step = function(index) {
+      const step = function (index) {
         const hookFn = queue[index];
 
         if (index >= queue.length) {
           next(data);
         } else if (typeof hookFn === 'function') {
+          const errTitle = 'Docsify plugin error';
+
           if (hookFn.length === 2) {
-            hookFn(data, result => {
-              data = result;
+            try {
+              hookFn(data, result => {
+                data = result;
+                step(index + 1);
+              });
+            } catch (err) {
+              if (catchPluginErrors) {
+                // eslint-disable-next-line no-console
+                console.error(errTitle, err);
+              } else {
+                throw err;
+              }
+
               step(index + 1);
-            });
+            }
           } else {
-            const result = hookFn(data);
-            data = result === undefined ? data : result;
-            step(index + 1);
+            try {
+              const result = hookFn(data);
+
+              data = result === undefined ? data : result;
+              step(index + 1);
+            } catch (err) {
+              if (catchPluginErrors) {
+                // eslint-disable-next-line no-console
+                console.error(errTitle, err);
+              } else {
+                throw err;
+              }
+
+              step(index + 1);
+            }
           }
         } else {
           step(index + 1);
