@@ -1,61 +1,21 @@
-import marked from 'marked';
-import { isAbsolutePath, getPath, getParentPath } from '../router/util';
-import { isFn, merge, cached, isPrimitive } from '../util/core';
-import { tree as treeTpl } from './tpl';
-import { genTree } from './gen-tree';
-import { slugify } from './slugify';
-import { emojify } from './emojify';
-import { getAndRemoveConfig, removeAtag } from './utils';
-import { imageCompiler } from './compiler/image';
-import { highlightCodeCompiler } from './compiler/code';
-import { paragraphCompiler } from './compiler/paragraph';
-import { taskListCompiler } from './compiler/taskList';
-import { taskListItemCompiler } from './compiler/taskListItem';
-import { linkCompiler } from './compiler/link';
+import { marked } from 'marked';
+import { isAbsolutePath, getPath, getParentPath } from '../router/util.js';
+import { isFn, cached, isPrimitive } from '../util/core.js';
+import { tree as treeTpl } from './tpl.js';
+import { genTree } from './gen-tree.js';
+import { slugify } from './slugify.js';
+import { emojify } from './emojify.js';
+import { getAndRemoveConfig } from './utils.js';
+import { imageCompiler } from './compiler/image.js';
+import { headingCompiler } from './compiler/heading.js';
+import { highlightCodeCompiler } from './compiler/code.js';
+import { paragraphCompiler } from './compiler/paragraph.js';
+import { taskListCompiler } from './compiler/taskList.js';
+import { taskListItemCompiler } from './compiler/taskListItem.js';
+import { linkCompiler } from './compiler/link.js';
+import { compileMedia } from './compiler/media.js';
 
 const cachedLinks = {};
-
-const compileMedia = {
-  markdown(url) {
-    return {
-      url,
-    };
-  },
-  mermaid(url) {
-    return {
-      url,
-    };
-  },
-  iframe(url, title) {
-    return {
-      html: `<iframe src="${url}" ${title ||
-        'width=100% height=400'}></iframe>`,
-    };
-  },
-  video(url, title) {
-    return {
-      html: `<video src="${url}" ${title || 'controls'}>Not Support</video>`,
-    };
-  },
-  audio(url, title) {
-    return {
-      html: `<audio src="${url}" ${title || 'controls'}>Not Support</audio>`,
-    };
-  },
-  code(url, title) {
-    let lang = url.match(/\.(\w+)$/);
-
-    lang = title || (lang && lang[1]);
-    if (lang === 'md') {
-      lang = 'markdown';
-    }
-
-    return {
-      url,
-      lang,
-    };
-  },
-};
 
 export class Compiler {
   constructor(config, router) {
@@ -69,18 +29,17 @@ export class Compiler {
       this.linkTarget === '_blank' ? config.externalLinkRel || 'noopener' : '';
     this.contentBase = router.getBasePath();
 
-    const renderer = this._initRenderer();
-    this.heading = renderer.heading;
+    this.renderer = this._initRenderer();
     let compile;
     const mdConf = config.markdown || {};
 
     if (isFn(mdConf)) {
-      compile = mdConf(marked, renderer);
+      compile = mdConf(marked, this.renderer);
     } else {
       marked.setOptions(
-        merge(mdConf, {
-          renderer: merge(renderer, mdConf.renderer),
-        })
+        Object.assign(mdConf, {
+          renderer: Object.assign(this.renderer, mdConf.renderer),
+        }),
       );
       compile = marked;
     }
@@ -88,7 +47,8 @@ export class Compiler {
     this._marked = compile;
     this.compile = text => {
       let isCached = true;
-      // eslint-disable-next-line no-unused-vars
+
+      // FIXME: this is not cached.
       const result = cached(_ => {
         isCached = false;
         let html = '';
@@ -103,7 +63,7 @@ export class Compiler {
           html = compile.parser(text);
         }
 
-        html = config.noEmoji ? html : emojify(html);
+        html = config.noEmoji ? html : emojify(html, config.nativeEmoji);
         slugify.clear();
 
         return html;
@@ -144,9 +104,9 @@ export class Compiler {
     if (config.include) {
       if (!isAbsolutePath(href)) {
         href = getPath(
-          process.env.SSR ? '' : this.contentBase,
+          this.contentBase,
           getParentPath(this.router.getCurrentPath()),
-          href
+          href,
         );
       }
 
@@ -168,7 +128,7 @@ export class Compiler {
           type = 'audio';
         }
 
-        embed = compileMedia[type].call(this, href, title);
+        embed = compileMedia[type](href, title);
         embed.type = type;
       }
 
@@ -181,8 +141,7 @@ export class Compiler {
   _matchNotCompileLink(link) {
     const links = this.config.noCompileLinks || [];
 
-    for (let i = 0; i < links.length; i++) {
-      const n = links[i];
+    for (const n of links) {
       const re = cachedLinks[n] || (cachedLinks[n] = new RegExp(`^${n}$`));
 
       if (re.test(link)) {
@@ -194,59 +153,22 @@ export class Compiler {
   _initRenderer() {
     const renderer = new marked.Renderer();
     const { linkTarget, linkRel, router, contentBase } = this;
-    const _self = this;
+    // Supports mermaid
     const origin = {};
 
-    /**
-     * Render anchor tag
-     * @link https://github.com/markedjs/marked#overriding-renderer-methods
-     * @param {String} text Text content
-     * @param {Number} level Type of heading (h<level> tag)
-     * @returns {String} Heading element
-     */
-    origin.heading = renderer.heading = function(text, level) {
-      let { str, config } = getAndRemoveConfig(text);
-      const nextToc = { level, title: removeAtag(str) };
-
-      if (/<!-- {docsify-ignore} -->/g.test(str)) {
-        str = str.replace('<!-- {docsify-ignore} -->', '');
-        nextToc.title = removeAtag(str);
-        nextToc.ignoreSubHeading = true;
-      }
-
-      if (/{docsify-ignore}/g.test(str)) {
-        str = str.replace('{docsify-ignore}', '');
-        nextToc.title = removeAtag(str);
-        nextToc.ignoreSubHeading = true;
-      }
-
-      if (/<!-- {docsify-ignore-all} -->/g.test(str)) {
-        str = str.replace('<!-- {docsify-ignore-all} -->', '');
-        nextToc.title = removeAtag(str);
-        nextToc.ignoreAllSubs = true;
-      }
-
-      if (/{docsify-ignore-all}/g.test(str)) {
-        str = str.replace('{docsify-ignore-all}', '');
-        nextToc.title = removeAtag(str);
-        nextToc.ignoreAllSubs = true;
-      }
-
-      const slug = slugify(config.id || str);
-      const url = router.toURL(router.getCurrentPath(), { id: slug });
-      nextToc.slug = url;
-      _self.toc.push(nextToc);
-
-      return `<h${level} id="${slug}"><a href="${url}" data-id="${slug}" class="anchor"><span>${str}</span></a></h${level}>`;
-    };
-
+    // renderer customizers
+    origin.heading = headingCompiler({
+      renderer,
+      router,
+      compiler: this,
+    });
     origin.code = highlightCodeCompiler({ renderer });
     origin.link = linkCompiler({
       renderer,
       router,
       linkTarget,
       linkRel,
-      compilerClass: _self,
+      compiler: this,
     });
     origin.paragraph = paragraphCompiler({ renderer });
     origin.image = imageCompiler({ renderer, contentBase, router });
@@ -259,8 +181,8 @@ export class Compiler {
   }
 
   /**
-   * Compile sidebar
-   * @param {String} text Text content
+   * Compile sidebar, it uses _sidebar.md (or specific file) or the content's headings toc to render sidebar.
+   * @param {String} text Text content from the sidebar file, maybe empty
    * @param {Number} level Type of heading (h<level> tag)
    * @returns {String} Sidebar element
    */
@@ -269,32 +191,39 @@ export class Compiler {
     const currentPath = this.router.getCurrentPath();
     let html = '';
 
+    // compile sidebar from _sidebar.md
     if (text) {
-      html = this.compile(text);
-    } else {
-      for (let i = 0; i < toc.length; i++) {
-        if (toc[i].ignoreSubHeading) {
-          const deletedHeaderLevel = toc[i].level;
-          toc.splice(i, 1);
-          // Remove headers who are under current header
-          for (
-            let j = i;
-            j < toc.length && deletedHeaderLevel < toc[j].level;
-            j++
-          ) {
-            toc.splice(j, 1) && j-- && i++;
-          }
-
-          i--;
+      return this.compile(text);
+    }
+    // compile sidebar from content's headings toc
+    for (let i = 0; i < toc.length; i++) {
+      if (toc[i].ignoreSubHeading) {
+        const deletedHeaderLevel = toc[i].depth;
+        toc.splice(i, 1);
+        // Remove headers who are under current header
+        for (
+          let j = i;
+          j < toc.length && deletedHeaderLevel < toc[j].depth;
+          j++
+        ) {
+          toc.splice(j, 1) && j-- && i++;
         }
-      }
 
-      const tree = this.cacheTree[currentPath] || genTree(toc, level);
-      html = treeTpl(tree, '<ul>{inner}</ul>');
-      this.cacheTree[currentPath] = tree;
+        i--;
+      }
     }
 
+    const tree = this.cacheTree[currentPath] || genTree(toc, level);
+    html = treeTpl(tree);
+    this.cacheTree[currentPath] = tree;
     return html;
+  }
+
+  /**
+   * When current content redirect to a new path file, clean pre content headings toc
+   */
+  resetToc() {
+    this.toc = [];
   }
 
   /**
@@ -303,16 +232,12 @@ export class Compiler {
    * @returns {String} Sub-sidebar element
    */
   subSidebar(level) {
-    if (!level) {
-      this.toc = [];
-      return;
-    }
-
     const currentPath = this.router.getCurrentPath();
     const { cacheTree, toc } = this;
 
     toc[0] && toc[0].ignoreAllSubs && toc.splice(0);
-    toc[0] && toc[0].level === 1 && toc.shift();
+    // remove the first heading from the toc if it is a top-level heading
+    toc[0] && toc[0].depth === 1 && toc.shift();
 
     for (let i = 0; i < toc.length; i++) {
       toc[i].ignoreSubHeading && toc.splice(i, 1) && i--;
@@ -325,12 +250,21 @@ export class Compiler {
     return treeTpl(tree);
   }
 
+  /**
+   * Compile the text to generate HTML heading element based on the level
+   * @param {*} text Text content, for now it is only from the _sidebar.md file
+   * @param {*} level Type of heading (h<level> tag), for now it is always 1
+   * @returns
+   */
   header(text, level) {
-    return this.heading(text, level);
-  }
-
-  article(text) {
-    return this.compile(text);
+    const tokenHeading = {
+      type: 'heading',
+      raw: text,
+      depth: level,
+      text: text,
+      tokens: [{ type: 'text', raw: text, text: text }],
+    };
+    return this.renderer.heading(tokenHeading);
   }
 
   /**
