@@ -1,17 +1,15 @@
-import { noop } from '../../util/core';
-import { on } from '../../util/dom';
-import { parseQuery, cleanPath, replaceSlug, endsWith } from '../util';
-import { History } from './base';
+import { isExternal, noop } from '../../util/core.js';
+import { on } from '../../util/dom.js';
+import { parseQuery, cleanPath, replaceSlug } from '../util.js';
+import { History } from './base.js';
 
 function replaceHash(path) {
   const i = location.href.indexOf('#');
   location.replace(location.href.slice(0, i >= 0 ? i : 0) + '#' + path);
 }
+
 export class HashHistory extends History {
-  constructor(config) {
-    super(config);
-    this.mode = 'hash';
-  }
+  mode = 'hash';
 
   getBasePath() {
     const path = window.location.pathname || '';
@@ -22,7 +20,7 @@ export class HashHistory extends History {
     // prevents the `/index.html` part of the URI from being
     // remove during routing.
     // See here: https://github.com/docsifyjs/docsify/pull/1372
-    const basePath = endsWith(path, '.html')
+    const basePath = path.endsWith('.html')
       ? path + '#/' + base
       : path + '/' + base;
     return /^(\/|https?:)/g.test(base) ? base : cleanPath(basePath);
@@ -47,7 +45,7 @@ export class HashHistory extends History {
     on('click', e => {
       const el = e.target.tagName === 'A' ? e.target : e.target.parentNode;
 
-      if (el && el.tagName === 'A' && !/_blank/.test(el.target)) {
+      if (el && el.tagName === 'A' && !isExternal(el.href)) {
         navigating = true;
       }
     });
@@ -73,7 +71,7 @@ export class HashHistory extends History {
 
   /**
    * Parse the url
-   * @param {string} [path=location.herf] URL to be parsed
+   * @param {string} [path=location.href] URL to be parsed
    * @return {object} { path, query }
    */
   parse(path = location.href) {
@@ -94,6 +92,7 @@ export class HashHistory extends History {
       path,
       file: this.getFile(path, true),
       query: parseQuery(query),
+      response: {},
     };
   }
 
