@@ -1,4 +1,3 @@
-/* eslint-disable no-unused-vars */
 const fixedPath = location.href.replace('/-/', '/#/');
 if (fixedPath !== location.href) {
   location.href = fixedPath;
@@ -27,8 +26,7 @@ function install(hook, vm) {
     div.style = `width: ${main.clientWidth}px; margin: 0 auto 20px;`;
     dom.appendTo(dom.find('.content'), div);
 
-    // eslint-disable-next-line
-    window.disqus_config = function() {
+    window.disqus_config = function () {
       this.page.url = location.origin + '/-' + vm.route.path;
       this.page.identifier = vm.route.path;
       this.page.title = document.title;
@@ -39,7 +37,7 @@ function install(hook, vm) {
     if (typeof window.DISQUS !== 'undefined') {
       window.DISQUS.reset({
         reload: true,
-        config: function() {
+        config() {
           this.page.url = location.origin + '/-' + vm.route.path;
           this.page.identifier = vm.route.path;
           this.page.title = document.title;
@@ -49,4 +47,5 @@ function install(hook, vm) {
   });
 }
 
-$docsify.plugins = [].concat(install, $docsify.plugins);
+window.$docsify = window.$docsify || {};
+$docsify.plugins = [install, ...($docsify.plugins || [])];
