@@ -1,7 +1,35 @@
-/* eslint-disable no-unused-vars */
-import { getAndRemoveConfig } from '../../core/render/utils';
+import {
+  getAndRemoveConfig,
+  getAndRemoveDocsifyIgnoreConfig,
+} from '../../core/render/utils.js';
+import { markdownToTxt } from './markdown-to-txt.js';
+import Dexie from 'dexie';
 
-let INDEXS = {};
+let INDEXES = {};
+
+const db = new Dexie('docsify');
+db.version(1).stores({
+  search: 'slug, title, body, path, indexKey',
+  expires: 'key, value',
+});
+
+async function saveData(maxAge, expireKey) {
+  INDEXES = Object.values(INDEXES).flatMap(innerData =>
+    Object.values(innerData),
+  );
+  await db.search.bulkPut(INDEXES);
+  await db.expires.put({ key: expireKey, value: Date.now() + maxAge });
+}
+
+async function getData(key, isExpireKey = false) {
+  if (isExpireKey) {
+    const item = await db.expires.get(key);
+    return item ? item.value : 0;
+  }
+
+  const item = await db.search.where({ indexKey: key }).toArray();
+  return item ? item : null;
+}
 
 const LOCAL_STORAGE = {
   EXPIRE_KEY: 'docsify.search.expires',
@@ -56,11 +84,9 @@ function getAllPaths(router) {
 
 function getTableData(token) {
   if (!token.text && token.type === 'table') {
-    token.cells.unshift(token.header);
-    token.text = token.cells
-      .map(function(rows) {
-        return rows.join(' | ');
-      })
+    token.rows.unshift(token.header);
+    token.text = token.rows
+      .map(columns => columns.map(r => r.text).join(' | '))
       .join(' |\n ');
   }
   return token.text;
@@ -73,45 +99,45 @@ function getListData(token) {
   return token.text;
 }
 
-function saveData(maxAge, expireKey, indexKey) {
-  localStorage.setItem(expireKey, Date.now() + maxAge);
-  localStorage.setItem(indexKey, JSON.stringify(INDEXS));
-}
-
-export function genIndex(path, content = '', router, depth) {
+export function genIndex(path, content = '', router, depth, indexKey) {
   const tokens = window.marked.lexer(content);
   const slugify = window.Docsify.slugify;
   const index = {};
   let slug;
   let title = '';
 
-  tokens.forEach(function(token, tokenIndex) {
+  tokens.forEach((token, tokenIndex) => {
     if (token.type === 'heading' && token.depth <= depth) {
       const { str, config } = getAndRemoveConfig(token.text);
+
+      const text = getAndRemoveDocsifyIgnoreConfig(token.text).content;
 
       if (config.id) {
         slug = router.toURL(path, { id: slugify(config.id) });
       } else {
-        slug = router.toURL(path, { id: slugify(escapeHtml(token.text)) });
+        slug = router.toURL(path, { id: slugify(escapeHtml(text)) });
       }
 
       if (str) {
-        title = str
-          .replace(/<!-- {docsify-ignore} -->/, '')
-          .replace(/{docsify-ignore}/, '')
-          .replace(/<!-- {docsify-ignore-all} -->/, '')
-          .replace(/{docsify-ignore-all}/, '')
-          .trim();
+        title = getAndRemoveDocsifyIgnoreConfig(str).content;
       }
 
-      index[slug] = { slug, title: title, body: '' };
+      index[slug] = {
+        slug,
+        title: title,
+        body: '',
+        path: path,
+        indexKey: indexKey,
+      };
     } else {
       if (tokenIndex === 0) {
         slug = router.toURL(path);
         index[slug] = {
           slug,
           title: path !== '/' ? path.slice(1) : 'Home Page',
-          body: token.text || '',
+          body: markdownToTxt(token.text || ''),
+          path: path,
+          indexKey: indexKey,
         };
       }
 
@@ -125,15 +151,16 @@ export function genIndex(path, content = '', router, depth) {
         token.text = getTableData(token);
         token.text = getListData(token);
 
-        index[slug].body += '\n' + (token.text || '');
+        index[slug].body += '\n' + markdownToTxt(token.text || '');
       } else {
         token.text = getTableData(token);
         token.text = getListData(token);
 
-        index[slug].body = index[slug].body
-          ? index[slug].body + token.text
-          : token.text;
+        index[slug].body = markdownToTxt(token.text || '');
       }
+
+      index[slug].path = path;
+      index[slug].indexKey = indexKey;
     }
   });
   slugify.clear();
@@ -153,19 +180,14 @@ export function ignoreDiacriticalMarks(keyword) {
  */
 export function search(query) {
   const matchingResults = [];
-  let data = [];
-  Object.keys(INDEXS).forEach(key => {
-    data = data.concat(Object.keys(INDEXS[key]).map(page => INDEXS[key][page]));
-  });
 
   query = query.trim();
   let keywords = query.split(/[\s\-，\\/]+/);
   if (keywords.length !== 1) {
-    keywords = [].concat(query, keywords);
+    keywords = [query, ...keywords];
   }
 
-  for (let i = 0; i < data.length; i++) {
-    const post = data[i];
+  for (const post of INDEXES) {
     let matchesScore = 0;
     let resultStr = '';
     let handlePostTitle = '';
@@ -180,9 +202,9 @@ export function search(query) {
         const regEx = new RegExp(
           escapeHtml(ignoreDiacriticalMarks(keyword)).replace(
             /[|\\{}()[\]^$+*?.]/g,
-            '\\$&'
+            '\\$&',
           ),
-          'gi'
+          'gi',
         );
         let indexTitle = -1;
         let indexContent = -1;
@@ -206,21 +228,17 @@ export function search(query) {
           let end = 0;
 
           start = indexContent < 11 ? 0 : indexContent - 10;
-          end = start === 0 ? 70 : indexContent + keyword.length + 60;
+          end = start === 0 ? 100 : indexContent + keyword.length + 90;
 
-          if (postContent && end > postContent.length) {
-            end = postContent.length;
+          if (handlePostContent && end > handlePostContent.length) {
+            end = handlePostContent.length;
           }
 
           const matchContent =
-            '...' +
+            handlePostContent &&
             handlePostContent
               .substring(start, end)
-              .replace(
-                regEx,
-                word => `<em class="search-keyword">${word}</em>`
-              ) +
-            '...';
+              .replace(regEx, word => /* html */ `<mark>${word}</mark>`);
 
           resultStr += matchContent;
         }
@@ -242,7 +260,7 @@ export function search(query) {
   return matchingResults.sort((r1, r2) => r2.score - r1.score);
 }
 
-export function init(config, vm) {
+export async function init(config, vm) {
   const isAuto = config.paths === 'auto';
   const paths = isAuto ? getAllPaths(vm.router) : config.paths;
 
@@ -255,7 +273,7 @@ export function init(config, vm) {
     if (Array.isArray(config.pathNamespaces)) {
       namespaceSuffix =
         config.pathNamespaces.filter(
-          prefix => path.slice(0, prefix.length) === prefix
+          prefix => path.slice(0, prefix.length) === prefix,
         )[0] || namespaceSuffix;
     } else if (config.pathNamespaces instanceof RegExp) {
       const matches = path.match(config.pathNamespaces);
@@ -276,12 +294,12 @@ export function init(config, vm) {
   const expireKey = resolveExpireKey(config.namespace) + namespaceSuffix;
   const indexKey = resolveIndexKey(config.namespace) + namespaceSuffix;
 
-  const isExpired = localStorage.getItem(expireKey) < Date.now();
+  const isExpired = (await getData(expireKey, true)) < Date.now();
 
-  INDEXS = JSON.parse(localStorage.getItem(indexKey));
+  INDEXES = await getData(indexKey);
 
   if (isExpired) {
-    INDEXS = {};
+    INDEXES = {};
   } else if (!isAuto) {
     return;
   }
@@ -290,15 +308,26 @@ export function init(config, vm) {
   let count = 0;
 
   paths.forEach(path => {
-    if (INDEXS[path]) {
+    const pathExists = Array.isArray(INDEXES)
+      ? INDEXES.some(obj => obj.path === path)
+      : false;
+    if (pathExists) {
       return count++;
     }
 
     Docsify.get(vm.router.getFile(path), false, vm.config.requestHeaders).then(
-      result => {
-        INDEXS[path] = genIndex(path, result, vm.router, config.depth);
-        len === ++count && saveData(config.maxAge, expireKey, indexKey);
-      }
+      async result => {
+        INDEXES[path] = genIndex(
+          path,
+          result,
+          vm.router,
+          config.depth,
+          indexKey,
+        );
+        if (len === ++count) {
+          await saveData(config.maxAge, expireKey);
+        }
+      },
     );
   });
 }
