@@ -1,30 +1,28 @@
-/* eslint-disable no-unused-vars */
 import mediumZoom from 'medium-zoom';
-
-const matchesSelector =
-  Element.prototype.matches ||
-  Element.prototype.webkitMatchesSelector ||
-  Element.prototype.msMatchesSelector;
 
 function install(hook) {
   let zoom;
 
   hook.doneEach(_ => {
-    let elms = Array.apply(
-      null,
+    let elms = Array.from(
       document.querySelectorAll(
-        '.markdown-section img:not(.emoji):not([data-no-zoom])'
-      )
+        '.markdown-section img:not(.emoji):not([data-no-zoom])',
+      ),
     );
 
-    elms = elms.filter(elm => matchesSelector.call(elm, 'a img') === false);
+    Docsify.dom.style(
+      `.medium-zoom-image--opened,.medium-zoom-overlay{z-index:999}`,
+    );
+
+    elms = elms.filter(elm => !elm.matches('a img'));
 
     if (zoom) {
       zoom.detach();
     }
 
-    zoom = mediumZoom(elms);
+    zoom = mediumZoom(elms, { background: 'var(--color-bg)' });
   });
 }
 
-$docsify.plugins = [].concat(install, $docsify.plugins);
+window.$docsify = window.$docsify || {};
+$docsify.plugins = [install, ...($docsify.plugins || [])];
